+++ examples/ts/client_side_intergation/client/src/app/colors.ts
@@ -1,0 +1,54 @@
+/* tslint:disable max-line-length */
+export const SLATE = {
+  50: '#F9FAFC',
+  100: '#F1F5F9',
+  200: '#E1E7F0',
+  300: '#CAD5E0',
+  400: '#95A2B7',
+  500: '#63738A',
+  600: '#475468',
+  700: '#334054',
+  800: '#1E2939',
+  900: '#0F1729',
+};
+
+// These will be getting updated and expanded
+export const INWORLD_COLORS = {
+  warmGray: {
+    10: '#FAF8F5',
+    20: '#EFEDEC',
+    30: '#D8D3D1',
+    50: '#96908E',
+    70: '#555151',
+    80: '#3B3838',
+  },
+  magenta: {
+    50: '#C333DB',
+    60: '#A719C2',
+    30: '#DE7DF0',
+    20: '#ECD8F0',
+  },
+  purple: {
+    10: '#FCFAFF',
+    20: '#EEE5FF',
+    40: '#A57CFF',
+    50: '#864CFF',
+    60: '#561EE5',
+    70: '#3000B2',
+    // 90: '#312352',
+    90: '#391C54',
+    100: '#140734',
+  },
+  green: {
+    50: '#1DDBA1',
+    60: '#19C28C',
+    40: '#7AEBC9',
+    20: '#DCF5ED',
+  },
+  orange: {
+    50: '#FF8766',
+    60: '#E04F28',
+    40: '#FFB894',
+    20: '#FFEFE7',
+  },
+};