+++ src/clients/inworld.client.ts
@@ -1,4 +1,3 @@
-import util = require('node:util');
 import { ServiceError } from '@grpc/grpc-js';
 import {
   CapabilitiesRequest,
@@ -103,12 +102,6 @@ export class InworldClient<
     return this;
   }
 
-  setGenerateSessionToken(generateSessionToken: GenerateSessionTokenFn) {
-    this.generateSessionTokenFn = generateSessionToken;
-
-    return this;
-  }
-
   setOnSession(props: GetterSetter<Session>) {
     this.sessionGetterSetter = props;
 
@@ -205,8 +198,3 @@ export class InworldClient<
     this.logger.error(err);
   };
 }
-
-InworldClient.prototype.setGenerateSessionToken = util.deprecate(
-  InworldClient.prototype.setGenerateSessionToken,
-  'setGenerateSessionToken() is deprecated. Use setOnSession() instead to manage session.',
-);